move block logic into blocks, fix broken favicons/logos
