+++ pages/blocks/background/background.js
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2021 Adobe. All rights reserved.
+ * This file is licensed to you under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License. You may obtain a copy
+ * of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under
+ * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+ * OF ANY KIND, either express or implied. See the License for the specific language
+ * governing permissions and limitations under the License.
+ */
+
+/** @type {import('../block.js').BlockDecorator} */
+export default async function decorate($block) {
+  const $images = $block.querySelectorAll('img');
+  const $lastImage = $images[$images.length - 1];
+
+  const $section = $block.closest('.section-wrapper');
+  if ($section && $lastImage) {
+    $section.style.backgroundImage = `url(${$lastImage.src})`;
+    let $caption = $lastImage.nextElementSibling;
+    if ($caption) {
+      if ($caption.textContent === '') $caption = $caption.nextElementSibling;
+      if ($caption) $caption.classList.add('background-image-caption');
+    }
+    $lastImage.remove();
+  }
+}
