+++ pages/blocks/linktexting/linktexting.js
@@ -12,20 +12,20 @@
 
 import { createTag } from '../../scripts/scripts.js';
 
-function decorateLinkTexting() {
-  const $linktexting = document.querySelector('main div.linktexting');
-  if ($linktexting) {
-    const $id = $linktexting.children[0].children[1].textContent;
-    // const $main = document.querySelector('main');
-    document.querySelectorAll('main p').forEach(($p) => {
-      if ($p.textContent.includes('<linktexting>')) {
-        // let $p=$p.parentElement;
-        const $widget = createTag('div', { id: 'linktexting-holder' });
-        $p.parentElement.replaceChild($widget, $p);
+function decorateLinkTexting($linktexting) {
+  // const $linktexting = document.querySelector('main div.linktexting');
+  // if ($linktexting) {
+  const $id = $linktexting.children[0].children[1].textContent;
+  // const $main = document.querySelector('main');
+  document.querySelectorAll('main p').forEach(($p) => {
+    if ($p.textContent.includes('<linktexting>')) {
+      // let $p=$p.parentElement;
+      const $widget = createTag('div', { id: 'linktexting-holder' });
+      $p.parentElement.replaceChild($widget, $p);
 
-        const parent = document.getElementById('linktexting-holder');
+      const parent = document.getElementById('linktexting-holder');
 
-        parent.innerHTML = `
+      parent.innerHTML = `
         <div class="promptWrapper">
           <div class="linkTextingWidgetWrapper" style="">
             <div class="linkTextingWidget" style="">
@@ -43,8 +43,8 @@ function decorateLinkTexting() {
         </div>
         `;
 
-        const $newHeadJS1 = createTag('script', { type: 'text/javascript' });
-        $newHeadJS1.innerHTML = `
+      const $newHeadJS1 = createTag('script', { type: 'text/javascript' });
+      $newHeadJS1.innerHTML = `
   /* Default Country Off, if set to false will use linkTextingDefaultCountry to decide default country, otherwise will determine the visiting user's country via their location */
   var linkTextingDefaultCountryOff = true;
   /* All Country data is lowercase ISO2 standard http://en.wikipedia.org/wiki/ISO_3166-1_alpha-2 */
@@ -58,18 +58,18 @@ function decorateLinkTexting() {
   var linkTextingOnlyCountries = ["af","al","dz","as","ad","ao","ai","ag","ar","am","aw","au","at","az","bs","bh","bd","bb","by","be","bz","bj","bm","bt","bo","ba","bw","br","io","vg","bn","bg","bf","bi","kh","cm","ca","cv","bq","ky","cf","td","cl","cn","co","km","cd","cg","ck","cr","ci","hr","cu","cw","cy","cz","dk","dj","dm","do","ec","eg","sv","gq","er","ee","et","fk","fo","fj","fi","fr","gf","pf","ga","gm","ge","de","gh","gi","gr","gl","gd","gp","gu","gt","gn","gw","gy","ht","hn","hk","hu","is","in","id","ir","iq","ie","il","it","jm","jp","jo","kz","ke","ki","kw","kg","la","lv","lb","ls","lr","ly","li","lt","lu","mo","mk","mg","mw","my","mv","ml","mt","mh","mq","mr","mu","mx","fm","md","mc","mn","me","ms","ma","mz","mm","na","nr","np","nl","nc","nz","ni","ne","ng","nu","nf","kp","mp","no","om","pk","pw","ps","pa","pg","py","pe","ph","pl","pt","pr","qa","re","ro","ru","rw","bl","sh","kn","lc","mf","pm","vc","ws","sm","st","sa","sn","rs","sc","sl","sg","sx","sk","si","sb","so","za","kr","ss","es","lk","sd","sr","sz","se","ch","sy","tw","tj","tz","th","tl","tg","tk","to","tt","tn","tr","tm","tc","tv","vi","ug","ua","ae","gb","us","uy","uz","vu","va","ve","vn","wf","ye","zm","zw"];
   `;
 
-        const $newHeadJS2 = createTag('script', { type: 'text/javascript', src: '//s3.amazonaws.com/linktexting-cdn/1.7/js/link_texting_gz.min.js' });
+      const $newHeadJS2 = createTag('script', { type: 'text/javascript', src: '//s3.amazonaws.com/linktexting-cdn/1.7/js/link_texting_gz.min.js' });
 
-        document.head.appendChild($newHeadJS1);
-        document.head.appendChild($newHeadJS2);
-      }
-    });
+      document.head.appendChild($newHeadJS1);
+      document.head.appendChild($newHeadJS2);
+    }
+  });
 
-    $linktexting.remove();
-  }
+  $linktexting.remove();
+  // }
 }
 
-/** @type {import('../block.js').BlockDecorator} */
+/** @type {import('../block').BlockDecorator} */
 export default function decorate($block) {
   decorateLinkTexting($block);
 }
