+++ pages/scripts/default-blocks.js
@@ -14,8 +14,8 @@
 
 import {
   createTag,
-  loadCSS,
-  loadLocalHeader,
+  // loadCSS,
+  // loadLocalHeader,
   toClassName,
 } from './scripts.js';
 
@@ -41,72 +41,72 @@ export function contentAsset() {
   });
 }
 
-function getImageName(appName) {
-  let iconName = '';
-  if (appName.toLowerCase().includes('adobe')) {
-    iconName = appName.toLowerCase().split('adobe')[1];
-  } else {
-    iconName = appName.toLowerCase();
-  }
-  return `/icons/${iconName.split(' ').join('')}`;
-}
+// function getImageName(appName) {
+//   let iconName = '';
+//   if (appName.toLowerCase().includes('adobe')) {
+//     iconName = appName.toLowerCase().split('adobe')[1];
+//   } else {
+//     iconName = appName.toLowerCase();
+//   }
+//   return `/icons/${iconName.split(' ').join('')}`;
+// }
 
-export function styleNav() {
-  const parent = document.querySelector('header');
-  const $appIcon = parent.querySelector('img');
-  if (!$appIcon) return;
-  const appIcon = $appIcon.src;
-  const appName = parent.querySelector('a').innerHTML;
-  const appNameLink = parent.querySelector('a').getAttribute('href');
-  const listItems = parent.querySelectorAll('ul li');
-  let nav = '';
-  let carrot = '';
+// export function styleNav() {
+//   const parent = document.querySelector('header');
+//   const $appIcon = parent.querySelector('img');
+//   if (!$appIcon) return;
+//   const appIcon = $appIcon.src;
+//   const appName = parent.querySelector('a').innerHTML;
+//   const appNameLink = parent.querySelector('a').getAttribute('href');
+//   const listItems = parent.querySelectorAll('ul li');
+//   let nav = '';
+//   let carrot = '';
 
-  if (listItems) {
-    if (listItems.length >= 1) {
-      nav = parent.querySelector('ul').outerHTML;
-      carrot = `
-        <div class="menu-carrot">
-          <img src='/icons/carrot.svg'>
-        </div>
-      `;
-    }
-  }
+//   if (listItems) {
+//     if (listItems.length >= 1) {
+//       nav = parent.querySelector('ul').outerHTML;
+//       carrot = `
+//         <div class="menu-carrot">
+//           <img src='/icons/carrot.svg'>
+//         </div>
+//       `;
+//     }
+//   }
 
-  parent.innerHTML = `
-    <div class="section-wrapper">
-      <div class="nav">
-        <div class="nav__section">
-          <div class="app-name-and-icon">
-            <div class="app-icon mobile"><img src="${appIcon}" alt="${appName}"></div>
-            <div class="app-icon desktop">
-              <a href="${appNameLink}" target="_blank">
-                <img src="${getImageName(appName)}.svg" alt="${appName}">
-              </a>
-            </div>
-            <div class="app-name mobile">
-              ${appName}
-            </div>
-            <div class="app-name desktop">
-              <a href="${appNameLink}" target="_blank">${appName}</a>
-            </div>
-            ${carrot}
-          </div>
-        </div> 
-        <nav>
-          ${nav}
-        </nav>
-      </div>
-    </div>
-  `;
+//   parent.innerHTML = `
+//     <div class="section-wrapper">
+//       <div class="nav">
+//         <div class="nav__section">
+//           <div class="app-name-and-icon">
+//             <div class="app-icon mobile"><img src="${appIcon}" alt="${appName}"></div>
+//             <div class="app-icon desktop">
+//               <a href="${appNameLink}" target="_blank">
+//                 <img src="${getImageName(appName)}.svg" alt="${appName}">
+//               </a>
+//             </div>
+//             <div class="app-name mobile">
+//               ${appName}
+//             </div>
+//             <div class="app-name desktop">
+//               <a href="${appNameLink}" target="_blank">${appName}</a>
+//             </div>
+//             ${carrot}
+//           </div>
+//         </div>
+//         <nav>
+//           ${nav}
+//         </nav>
+//       </div>
+//     </div>
+//   `;
 
-  if (document.querySelector('.nav-container')) {
-    const $nav = document.querySelector('.nav-container');
-    if ($nav.querySelector('div').children.length === 0) {
-      $nav.remove();
-    }
-  }
-}
+//   if (document.querySelector('.nav-container')) {
+//     const $nav = document.querySelector('.nav-container');
+//     if ($nav.querySelector('div').children.length === 0) {
+//       $nav.remove();
+//     }
+//   }
+// }
 
 export function setExternalLinks() {
   if (!document.querySelectorAll('main a')) return;
@@ -242,15 +242,15 @@ export function decorateNextStep() {
   `;
 }
 
-export function mobileDropDown() {
-  // event.preventDefault();
-  const body = document.getElementsByTagName('body')[0];
-  if (!body.classList.contains('nav-showing')) {
-    body.classList.add('nav-showing');
-  } else {
-    body.classList.remove('nav-showing');
-  }
-}
+// export function mobileDropDown() {
+//   // event.preventDefault();
+//   const body = document.getElementsByTagName('body')[0];
+//   if (!body.classList.contains('nav-showing')) {
+//     body.classList.add('nav-showing');
+//   } else {
+//     body.classList.remove('nav-showing');
+//   }
+// }
 
 export function decorateEmbeds() {
   document.querySelectorAll('a[href]').forEach(($a) => {
@@ -330,37 +330,37 @@ export function decorateTables() {
 //   return config;
 // }
 
-export function decorateBackgroundImageBlocks() {
-  document.querySelectorAll('main div.background-image').forEach(($bgImgDiv) => {
-    const $images = $bgImgDiv.querySelectorAll('img');
-    const $lastImage = $images[$images.length - 1];
+// export function decorateBackgroundImageBlocks() {
+//   document.querySelectorAll('main div.background-image').forEach(($bgImgDiv) => {
+//     const $images = $bgImgDiv.querySelectorAll('img');
+//     const $lastImage = $images[$images.length - 1];
 
-    const $section = $bgImgDiv.closest('.section-wrapper');
-    if ($section && $lastImage) {
-      $section.style.backgroundImage = `url(${$lastImage.src})`;
-      let $caption = $lastImage.nextElementSibling;
-      if ($caption) {
-        if ($caption.textContent === '') $caption = $caption.nextElementSibling;
-        if ($caption) $caption.classList.add('background-image-caption');
-      }
-      $lastImage.remove();
-    }
-  });
-}
+//     const $section = $bgImgDiv.closest('.section-wrapper');
+//     if ($section && $lastImage) {
+//       $section.style.backgroundImage = `url(${$lastImage.src})`;
+//       let $caption = $lastImage.nextElementSibling;
+//       if ($caption) {
+//         if ($caption.textContent === '') $caption = $caption.nextElementSibling;
+//         if ($caption) $caption.classList.add('background-image-caption');
+//       }
+//       $lastImage.remove();
+//     }
+//   });
+// }
 
-export async function decorateNav() {
-  await loadLocalHeader();
-  styleNav();
-  const navEl = document.querySelector('.nav');
-  if (navEl) {
-    const iconEl = document.querySelector('.app-name-and-icon');
-    if (iconEl) {
-      iconEl.addEventListener('click', mobileDropDown);
-    }
-  }
-  document.querySelector('header').classList.add('appear');
-  loadCSS('/pages/blocks/nav/nav.css');
-}
+// export async function decorateNav() {
+//   await loadLocalHeader();
+//   styleNav();
+//   const navEl = document.querySelector('.nav');
+//   if (navEl) {
+//     const iconEl = document.querySelector('.app-name-and-icon');
+//     if (iconEl) {
+//       iconEl.addEventListener('click', mobileDropDown);
+//     }
+//   }
+//   document.querySelector('header').classList.add('appear');
+//   loadCSS('/pages/blocks/nav/nav.css');
+// }
 
 export function decorateButtons() {
   document.querySelectorAll('main a').forEach(($a) => {
@@ -393,61 +393,4 @@ export function decorateVideoBlocks() {
     }
     $a.parentElement.replaceChild($video, $a);
   });
-}
-
-export function decorateLinkTexting() {
-  const $linktexting = document.querySelector('main div.linktexting');
-  if ($linktexting) {
-    const $id = $linktexting.children[0].children[1].textContent;
-    // const $main = document.querySelector('main');
-    document.querySelectorAll('main p').forEach(($p) => {
-      if ($p.textContent.includes('<linktexting>')) {
-        // let $p=$p.parentElement;
-        const $widget = createTag('div', { id: 'linktexting-holder' });
-        $p.parentElement.replaceChild($widget, $p);
-
-        const parent = document.getElementById('linktexting-holder');
-
-        parent.innerHTML = `
-        <div class="promptWrapper">
-          <div class="linkTextingWidgetWrapper" style="">
-            <div class="linkTextingWidget" style="">
-              <div class="promptContent" style=""></div>
-              <div class="linkTextingInner">
-                <input type="hidden" class="linkID" value="${$id}">
-                  <div class="linkTextingInputWrapper">
-                    <input class="linkTextingInput linkTextingInputFlagAdjust" type="tel" id="numberToText_linkTexting">
-                  </div>
-                  <button class="linkTextingButton localized-button localized-text-text_me_a_link" type="button" id="sendButton_linkTexting" style="background-color: #1473E6;color : #ffffff">Text me a link</button>
-                  <div class="linkTextingError" id="linkTextingError" style="display:none;"></div>
-              </div>
-            </div>
-          </div>
-        </div>
-        `;
-
-        const $newHeadJS1 = createTag('script', { type: 'text/javascript' });
-        $newHeadJS1.innerHTML = `
-  /* Default Country Off, if set to false will use linkTextingDefaultCountry to decide default country, otherwise will determine the visiting user's country via their location */
-  var linkTextingDefaultCountryOff = true;
-  /* All Country data is lowercase ISO2 standard http://en.wikipedia.org/wiki/ISO_3166-1_alpha-2 */
-  /* Default Country, if set to "auto" will determine the visiting user's country via their location */
-  var linkTextingDefaultCountry = "us";
-  /* If using automatic country detection with https or over 1,000 visits/day go to http://ipinfo.io, get a token, and place it here */
-  var linkTextingIpinfoToken = "";
-  /* Preferred Countries in Dropdown */
-  var linkTextingPreferredCountries = ["us","ca"];
-  /* Countries available in dropdown */
-  var linkTextingOnlyCountries = ["af","al","dz","as","ad","ao","ai","ag","ar","am","aw","au","at","az","bs","bh","bd","bb","by","be","bz","bj","bm","bt","bo","ba","bw","br","io","vg","bn","bg","bf","bi","kh","cm","ca","cv","bq","ky","cf","td","cl","cn","co","km","cd","cg","ck","cr","ci","hr","cu","cw","cy","cz","dk","dj","dm","do","ec","eg","sv","gq","er","ee","et","fk","fo","fj","fi","fr","gf","pf","ga","gm","ge","de","gh","gi","gr","gl","gd","gp","gu","gt","gn","gw","gy","ht","hn","hk","hu","is","in","id","ir","iq","ie","il","it","jm","jp","jo","kz","ke","ki","kw","kg","la","lv","lb","ls","lr","ly","li","lt","lu","mo","mk","mg","mw","my","mv","ml","mt","mh","mq","mr","mu","mx","fm","md","mc","mn","me","ms","ma","mz","mm","na","nr","np","nl","nc","nz","ni","ne","ng","nu","nf","kp","mp","no","om","pk","pw","ps","pa","pg","py","pe","ph","pl","pt","pr","qa","re","ro","ru","rw","bl","sh","kn","lc","mf","pm","vc","ws","sm","st","sa","sn","rs","sc","sl","sg","sx","sk","si","sb","so","za","kr","ss","es","lk","sd","sr","sz","se","ch","sy","tw","tj","tz","th","tl","tg","tk","to","tt","tn","tr","tm","tc","tv","vi","ug","ua","ae","gb","us","uy","uz","vu","va","ve","vn","wf","ye","zm","zw"];
-  `;
-
-        const $newHeadJS2 = createTag('script', { type: 'text/javascript', src: '//s3.amazonaws.com/linktexting-cdn/1.7/js/link_texting_gz.min.js' });
-
-        document.head.appendChild($newHeadJS1);
-        document.head.appendChild($newHeadJS2);
-      }
-    });
-
-    $linktexting.remove();
-  }
 }
